+++ src/Bundle/AutoMapperBundle/Tests/Fixtures/NestedObject.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Jane\Bundle\AutoMapperBundle\Tests\Fixtures;
+
+class NestedObject
+{
+    /**
+     * @var User
+     */
+    public $user;
+}